Reconfigured file path
