+++ front_end/src/app/pages/home/home.component.html
@@ -1,41 +1,57 @@
 <div class="home-page">
 
-<body>
-    <div class="content">
-        <h1>Welcome to the future of wireless communication! </h1>
-<p>
-    
-    At Spartan Wireless, we're revolutionizing the way you connect with the world. Say goodbye to dropped calls, slow data speeds, and restrictive contracts. With our cutting-edge wireless phone service, you'll experience seamless connectivity like never before.
-<br>
-<br></p>
-    <h3>Why choose us?</h3><p>
-    
-    <br>
-    <strong>Blazing Fast Speeds:</strong> Stream, download, and browse with lightning-fast 5G speeds. Whether you're streaming your favorite movies or video chatting with loved ones, you'll never experience lag or buffering again.
-    <br>
-    <br>
-    <strong>Nationwide Coverage:</strong> Stay connected wherever you go with our expansive nationwide coverage. Whether you're in the heart of the city or off the beaten path, our network has you covered.
-    <br>
-    <br>
+    <body>
+        <div class="content">
+            <h1>Welcome to the future of wireless communication! </h1>
+            <p>
 
-    <strong>Flexible Plans:</strong> We understand that one size doesn't fit all. That's why we offer flexible plans tailored to your needs. Whether you're a light user or a data-hungry streamer, we have a plan that fits your lifestyle and budget.
-    <br>
-    <br>
+                At Spartan Wireless, we're revolutionizing the way you connect with the world. Say goodbye to dropped
+                calls, slow data speeds, and restrictive contracts. With our cutting-edge wireless phone service, you'll
+                experience seamless connectivity like never before.
+                <br>
+                <br>
+            </p>
+            <h3>Why choose us?</h3>
+            <p>
 
-    <strong>No Contracts, No Hidden Fees:</strong> Say goodbye to long-term contracts and surprise fees. With us, what you see is what you get. Enjoy transparent pricing and the freedom to switch plans or cancel anytime without penalties.
-    <br>
-    <br>
+                <br>
+                <strong>Blazing Fast Speeds:</strong> Stream, download, and browse with lightning-fast 5G speeds.
+                Whether you're streaming your favorite movies or video chatting with loved ones, you'll never experience
+                lag or buffering again.
+                <br>
+                <br>
+                <strong>Nationwide Coverage:</strong> Stay connected wherever you go with our expansive nationwide
+                coverage. Whether you're in the heart of the city or off the beaten path, our network has you covered.
+                <br>
+                <br>
 
-    <strong>Exceptional Customer Service:</strong> Our dedicated team of customer service experts is here to support you every step of the way. Have a question or need assistance? We're just a call or click away, ready to provide you with top-notch service.
-    <br>
-    <br>
+                <strong>Flexible Plans:</strong> We understand that one size doesn't fit all. That's why we offer
+                flexible plans tailored to your needs. Whether you're a light user or a data-hungry streamer, we have a
+                plan that fits your lifestyle and budget.
+                <br>
+                <br>
 
-    <strong>Cutting-Edge Technology:</strong> We're constantly investing in the latest technology to ensure you have access to the best wireless experience possible. From advanced network infrastructure to innovative devices, we're committed to keeping you connected in the digital age.
-    
-    <br>
-    <br>
-    <b>Ready to experience the future of wireless communication? Join the Spartan Wireless family today and unlock a world of possibilities. Sign up now and take your connectivity to the next level!</b>
+                <strong>No Contracts, No Hidden Fees:</strong> Say goodbye to long-term contracts and surprise fees.
+                With us, what you see is what you get. Enjoy transparent pricing and the freedom to switch plans or
+                cancel anytime without penalties.
+                <br>
+                <br>
+
+                <strong>Exceptional Customer Service:</strong> Our dedicated team of customer service experts is here to
+                support you every step of the way. Have a question or need assistance? We're just a call or click away,
+                ready to provide you with top-notch service.
+                <br>
+                <br>
+
+                <strong>Cutting-Edge Technology:</strong> We're constantly investing in the latest technology to ensure
+                you have access to the best wireless experience possible. From advanced network infrastructure to
+                innovative devices, we're committed to keeping you connected in the digital age.
+
+                <br>
+                <br>
+                <b>Ready to experience the future of wireless communication? Join the Spartan Wireless family today and
+                    unlock a world of possibilities. Sign up now and take your connectivity to the next level!</b>
             </p>
         </div>
-    
-        </body>
+
+    </body>
